requirements.txt
customtkinter==5.2.2
pillow==10.2.0
pillow_heif==0.16.0
pyinstaller==6.3.0
